LPS-8302
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@48024 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/util/PropsValues.java
@@ -584,7 +584,7 @@ public class PropsValues {
 	public static final String[] DL_FILE_EXTENSIONS =
 		PropsUtil.getArray(PropsKeys.DL_FILE_EXTENSIONS);
 
-	public static final int DL_FILE_MAX_SIZE = GetterUtil.getInteger(
+	public static final long DL_FILE_MAX_SIZE = GetterUtil.getLong(
 		PropsUtil.get(PropsKeys.DL_FILE_MAX_SIZE));
 
 	public static final boolean DL_FILE_RANK_ENABLED = GetterUtil.getBoolean(
